Create navigating the project
Update README.md by adding a reference guide on how to use and access resources within the project on the website, and within the repositories. This should help to enable users to get a more intuitive feel as to how to access resources and what different terms or features mean. This is meant to also clear up some discrepancies between the project repository and where certain features are linked on the site.